IMPORTANT FIX: integer firing rate calculation - decoding precision FIX

+++ sdl_example/KDE_Estimator.cpp
@@ -372,7 +372,7 @@ int main(int argc, char **argv){
 
 	// compute occupancy KDE - pi(x) from tracking position sampling
 	// overall tetrode average firing rate, spikes / s
-	double mu = MIN_SPIKES * SAMPLING_RATE * BUFFER_SAMPLING_RATE / (BUFFER_LAST_PKG_ID - SAMPLING_DELAY);
+	double mu = MIN_SPIKES * SAMPLING_RATE * BUFFER_SAMPLING_RATE / double(BUFFER_LAST_PKG_ID - SAMPLING_DELAY);
 	std::cout << "t " << tetr << ": Average firing rate: " << mu << "\n";
 	for (int xb = 0; xb < NBINS; ++xb) {
 		for (int yb = 0; yb < NBINS; ++yb) {
